Phase 1: extract composer and landing modules from SidebarChat
Split VoidChatArea, SelectedFiles, ScrollToBottomContainer, and the landing page UI into dedicated modules to shrink SidebarChat and continue Sprint 1 modularization.
