+++ src/vs/workbench/contrib/cortexide/browser/react/src/sidebar-tsx/SidebarChat.tsx
@@ -3,7 +3,7 @@
  *  Licensed under the Apache License, Version 2.0. See LICENSE.txt for more information.
  *--------------------------------------------------------------------------------------*/
 
-import React, { ButtonHTMLAttributes, FormEvent, FormHTMLAttributes, Fragment, KeyboardEvent, useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import React, { FormEvent, FormHTMLAttributes, Fragment, KeyboardEvent, useCallback, useEffect, useMemo, useRef, useState } from 'react';
 
 
 import { useAccessor, useChatThreadsState, useChatThreadsStreamState, useSettingsState, useActiveURI, useCommandBarState, useFullChatThreadsStreamState } from '../util/services.js';
@@ -12,18 +12,21 @@ import { ChatMarkdownRender, ChatMessageLocation, getApplyBoxId } from '../markd
 import { URI } from '../../../../../../../base/common/uri.js';
 import { IDisposable } from '../../../../../../../base/common/lifecycle.js';
 import { ErrorDisplay } from './ErrorDisplay.js';
-import { BlockCode, TextAreaFns, VoidCustomDropdownBox, VoidInputBox2, VoidSlider, VoidSwitch, VoidDiffEditor } from '../util/inputs.js';
-import { ModelDropdown, } from '../settings/ModelDropdown.js';
+import { BlockCode, TextAreaFns, VoidInputBox2, VoidDiffEditor } from '../util/inputs.js';
 import { PastThreadsList } from './SidebarThreadSelector.js';
+import { VoidChatArea, ButtonSubmit, ButtonStop } from './composer/VoidChatArea.js';
+import { SelectedFiles } from './composer/SelectedFiles.js';
+import { ScrollToBottomContainer } from './composer/ScrollToBottomContainer.js';
+import { LandingPage } from './landing/LandingPage.js';
 import { ComposerTabs } from './chrome/ComposerTabs.js';
 import { ThreadHeader } from './chrome/ThreadHeader.js';
 import { CORTEXIDE_CTRL_L_ACTION_ID } from '../../../actionIDs.js';
 import { CORTEXIDE_OPEN_SETTINGS_ACTION_ID } from '../../../cortexideSettingsPane.js';
-import { ChatMode, displayInfoOfProviderName, FeatureName, isFeatureNameDisabled, isValidProviderModelSelection } from '../../../../../../../workbench/contrib/cortexide/common/cortexideSettingsTypes.js';
+import { displayInfoOfProviderName, FeatureName, isFeatureNameDisabled, isValidProviderModelSelection } from '../../../../../../../workbench/contrib/cortexide/common/cortexideSettingsTypes.js';
 import { ICommandService } from '../../../../../../../platform/commands/common/commands.js';
 import { WarningBox } from '../settings/WarningBox.js';
 import { getModelCapabilities, getIsReasoningEnabledState, getReservedOutputTokenSpace } from '../../../../common/modelCapabilities.js';
-import { AlertTriangle, File, Ban, Check, ChevronRight, Dot, FileIcon, Pencil, Undo, Undo2, X, Flag, Copy as CopyIcon, Info, CirclePlus, Ellipsis, CircleEllipsis, Folder, ALargeSmall, TypeOutline, Text, Image as ImageIcon, FileText } from 'lucide-react';
+import { AlertTriangle, File, Ban, Check, ChevronRight, Dot, FileIcon, Pencil, Undo, Undo2, X, Flag, Copy as CopyIcon, Info, CirclePlus, Ellipsis, CircleEllipsis, Folder, ALargeSmall, TypeOutline, Text } from 'lucide-react';
 import { ChatMessage, CheckpointEntry, StagingSelectionItem, ToolMessage, PlanMessage, ReviewMessage, PlanStep, StepStatus, PlanApprovalState } from '../../../../common/chatThreadServiceTypes.js';
 import { approvalTypeOfBuiltinToolName, BuiltinToolCallParams, BuiltinToolName, ToolName, LintErrorItem, ToolApprovalType, toolApprovalTypes } from '../../../../common/toolsServiceTypes.js';
 import { CopyButton, EditToolAcceptRejectButtonsHTML, IconShell1, JumpToFileButton, JumpToTerminalButton, StatusIndicator, StatusIndicatorForApplyButton, useApplyStreamState, useEditToolStreamState } from '../markdown/ApplyBlockHoverButtons.js';
@@ -39,13 +42,12 @@ import { removeMCPToolNamePrefix } from '../../../../common/mcpServiceTypes.js';
 import { useImageAttachments } from '../util/useImageAttachments.js';
 import { usePDFAttachments } from '../util/usePDFAttachments.js';
 import { useTranslation } from '../util/useTranslation.js';
-import { FileAccess } from '../../../../../../../base/common/network.js';
 import { PDFAttachmentList } from '../util/PDFAttachmentList.js';
 import { ImageAttachmentList } from '../util/ImageAttachmentList.js';
 import { ChatImageAttachment, ChatPDFAttachment } from '../../../../common/chatThreadServiceTypes.js';
 import { ImageMessageRenderer } from '../util/ImageMessageRenderer.js';
 import { PDFMessageRenderer } from '../util/PDFMessageRenderer.js';
-import { IconX, IconArrowUp, IconSquare, IconWarning, IconLoading, TypingCursor } from './shared/icons.js';
+import { IconX, IconWarning, IconLoading, TypingCursor } from './shared/icons.js';
 import { getBasename, getFolderName, getRelative, voidOpenFileFn } from './shared/pathUtils.js';
 import { ToolChildrenWrapper, CodeChildren, ListableToolItem } from './tools/ToolPrimitives.js';
 
@@ -53,775 +55,9 @@ import { ToolChildrenWrapper, CodeChildren, ListableToolItem } from './tools/Too
 export { IconX, IconWarning, IconLoading, TypingCursor } from './shared/icons.js';
 export { getBasename, getFolderName, getRelative, voidOpenFileFn } from './shared/pathUtils.js';
 export { ToolChildrenWrapper, CodeChildren, ListableToolItem } from './tools/ToolPrimitives.js';
+export { VoidChatArea, ButtonSubmit, ButtonStop } from './composer/VoidChatArea.js';
+export { SelectedFiles } from './composer/SelectedFiles.js';
 
-// SLIDER ONLY:
-const ReasoningOptionSlider = ({ featureName }: { featureName: FeatureName }) => {
-	const accessor = useAccessor()
-
-	const cortexideSettingsService = accessor.get('ICortexideSettingsService')
-	const voidSettingsState = useSettingsState()
-
-	const modelSelection = voidSettingsState.modelSelectionOfFeature[featureName]
-	const overridesOfModel = voidSettingsState.overridesOfModel
-
-	if (!modelSelection) return null
-
-	// Skip "auto" - it's not a real provider
-	if (!isValidProviderModelSelection(modelSelection)) {
-		return null;
-	}
-
-	const { modelName, providerName } = modelSelection
-	const { reasoningCapabilities } = getModelCapabilities(providerName, modelName, overridesOfModel)
-	const { canTurnOffReasoning, reasoningSlider: reasoningBudgetSlider } = reasoningCapabilities || {}
-
-	const modelSelectionOptions = voidSettingsState.optionsOfModelSelection[featureName][providerName]?.[modelName]
-	const isReasoningEnabled = getIsReasoningEnabledState(featureName, providerName, modelName, modelSelectionOptions, overridesOfModel)
-
-	if (canTurnOffReasoning && !reasoningBudgetSlider) { // if it's just a on/off toggle without a power slider
-		return <div className='flex items-center gap-x-2'>
-			<span className='text-void-fg-3 text-xs pointer-events-none inline-block w-10 pr-1'>Thinking</span>
-			<VoidSwitch
-				size='xxs'
-				value={isReasoningEnabled}
-				onChange={(newVal) => {
-					const isOff = canTurnOffReasoning && !newVal
-					cortexideSettingsService.setOptionsOfModelSelection(featureName, modelSelection.providerName, modelSelection.modelName, { reasoningEnabled: !isOff })
-				}}
-			/>
-		</div>
-	}
-
-	if (reasoningBudgetSlider?.type === 'budget_slider') { // if it's a slider
-		const { min: min_, max, default: defaultVal } = reasoningBudgetSlider
-
-		const nSteps = 8 // only used in calculating stepSize, stepSize is what actually matters
-		const stepSize = Math.round((max - min_) / nSteps)
-
-		const valueIfOff = min_ - stepSize
-		const min = canTurnOffReasoning ? valueIfOff : min_
-		const value = isReasoningEnabled ? voidSettingsState.optionsOfModelSelection[featureName][modelSelection.providerName]?.[modelSelection.modelName]?.reasoningBudget ?? defaultVal
-			: valueIfOff
-
-		return <div className='flex items-center gap-x-2'>
-			<span className='text-void-fg-3 text-xs pointer-events-none inline-block w-10 pr-1'>Thinking</span>
-			<VoidSlider
-				width={50}
-				size='xs'
-				min={min}
-				max={max}
-				step={stepSize}
-				value={value}
-				onChange={(newVal) => {
-					if (modelSelection.providerName === 'auto' && modelSelection.modelName === 'auto') return;
-					const isOff = canTurnOffReasoning && newVal === valueIfOff
-					cortexideSettingsService.setOptionsOfModelSelection(featureName, modelSelection.providerName, modelSelection.modelName, { reasoningEnabled: !isOff, reasoningBudget: newVal })
-				}}
-			/>
-			<span className='text-void-fg-3 text-xs pointer-events-none'>{isReasoningEnabled ? `${value} tokens` : 'Thinking disabled'}</span>
-		</div>
-	}
-
-	if (reasoningBudgetSlider?.type === 'effort_slider') {
-
-		const { values, default: defaultVal } = reasoningBudgetSlider
-
-		const min = canTurnOffReasoning ? -1 : 0
-		const max = values.length - 1
-
-		const currentEffort = voidSettingsState.optionsOfModelSelection[featureName][modelSelection.providerName]?.[modelSelection.modelName]?.reasoningEffort ?? defaultVal
-		const valueIfOff = -1
-		const value = isReasoningEnabled && currentEffort ? values.indexOf(currentEffort) : valueIfOff
-
-		const currentEffortCapitalized = currentEffort.charAt(0).toUpperCase() + currentEffort.slice(1, Infinity)
-
-		return <div className='flex items-center gap-x-2'>
-			<span className='text-void-fg-3 text-xs pointer-events-none inline-block w-10 pr-1'>Thinking</span>
-			<VoidSlider
-				width={30}
-				size='xs'
-				min={min}
-				max={max}
-				step={1}
-				value={value}
-				onChange={(newVal) => {
-					if (modelSelection.providerName === 'auto' && modelSelection.modelName === 'auto') return;
-					const isOff = canTurnOffReasoning && newVal === valueIfOff
-					cortexideSettingsService.setOptionsOfModelSelection(featureName, modelSelection.providerName, modelSelection.modelName, { reasoningEnabled: !isOff, reasoningEffort: values[newVal] ?? undefined })
-				}}
-			/>
-			<span className='text-void-fg-3 text-xs pointer-events-none'>{isReasoningEnabled ? `${currentEffortCapitalized}` : 'Thinking disabled'}</span>
-		</div>
-	}
-
-	return null
-}
-
-
-
-const nameOfChatMode: Record<ChatMode, string> = {
-	'normal': 'Ask',
-	'gather': 'Gather',
-	'plan': 'Plan',
-	'agent': 'Agent',
-}
-
-const detailOfChatMode: Record<ChatMode, string> = {
-	'normal': 'Ask questions about your code',
-	'gather': 'Reads files, no edits',
-	'plan': 'Plans first, then executes',
-	'agent': 'Edits files and uses tools',
-}
-
-
-const ChatModeDropdown = ({ className }: { className: string }) => {
-	const accessor = useAccessor()
-
-	const cortexideSettingsService = accessor.get('ICortexideSettingsService')
-	const settingsState = useSettingsState()
-
-	const options: ChatMode[] = useMemo(() => ['normal', 'gather', 'plan', 'agent'], [])
-
-	const onChangeOption = useCallback((newVal: ChatMode) => {
-		cortexideSettingsService.setGlobalSetting('chatMode', newVal)
-	}, [cortexideSettingsService])
-
-	const getModeDisplayName = (val: ChatMode) => nameOfChatMode[val] ?? val
-
-	return <VoidCustomDropdownBox
-		className={className}
-		options={options}
-		selectedOption={settingsState.globalSettings.chatMode}
-		onChangeOption={onChangeOption}
-		getOptionDisplayName={getModeDisplayName}
-		getOptionDropdownName={getModeDisplayName}
-		getOptionDropdownDetail={(val) => detailOfChatMode[val]}
-		getOptionsEqual={(a, b) => a === b}
-	/>
-
-}
-
-
-
-
-
-interface CortexideChatAreaProps {
-	// Required
-	children: React.ReactNode; // This will be the input component
-
-	// Form controls
-	onSubmit: () => void;
-	onAbort: () => void;
-	isStreaming: boolean;
-	isDisabled?: boolean;
-	divRef?: React.RefObject<HTMLDivElement | null>;
-
-	// UI customization
-	className?: string;
-	showModelDropdown?: boolean;
-	showSelections?: boolean;
-	showProspectiveSelections?: boolean;
-	loadingIcon?: React.ReactNode;
-
-	selections?: StagingSelectionItem[]
-	setSelections?: (s: StagingSelectionItem[]) => void
-	// selections?: any[];
-	// onSelectionsChange?: (selections: any[]) => void;
-
-	onClickAnywhere?: () => void;
-	// Optional close button
-	onClose?: () => void;
-
-	// Image attachments
-	imageAttachments?: React.ReactNode;
-	onImagePaste?: (files: File[]) => void;
-	onImageDrop?: (files: File[]) => void;
-	onImageUpload?: () => void;
-	onPDFDrop?: (files: File[]) => void;
-	pdfAttachments?: React.ReactNode;
-
-	featureName: FeatureName;
-}
-
-export const VoidChatArea: React.FC<CortexideChatAreaProps> = ({
-	children,
-	onSubmit,
-	onAbort,
-	onClose,
-	onClickAnywhere,
-	divRef,
-	isStreaming = false,
-	isDisabled = false,
-	className = '',
-	showModelDropdown = true,
-	showSelections = false,
-	showProspectiveSelections = false,
-	selections,
-	setSelections,
-	imageAttachments,
-	onImagePaste,
-	onImageDrop,
-	onImageUpload,
-	onPDFDrop,
-	pdfAttachments,
-	featureName,
-	loadingIcon,
-}) => {
-	const [isDragOver, setIsDragOver] = React.useState(false);
-	const imageInputRef = React.useRef<HTMLInputElement>(null);
-	const pdfInputRef = React.useRef<HTMLInputElement>(null);
-	const containerRef = React.useRef<HTMLDivElement>(null);
-
-		// allow-any-unicode-next-line
-		// Handle paste — listens on container (bubbles from textarea) AND document level
-	// (document listener catches Ctrl+V when the sidebar panel is focused but textarea isn't)
-	React.useEffect(() => {
-		const handlePaste = (e: ClipboardEvent) => {
-			const items = Array.from(e.clipboardData?.items || []);
-			const imageFiles: File[] = [];
-			const pdfFiles: File[] = [];
-
-			for (const item of items) {
-				if (item.type.startsWith('image/')) {
-					const file = item.getAsFile();
-					if (file) imageFiles.push(file);
-				} else if (item.type === 'application/pdf') {
-					const file = item.getAsFile();
-					if (file) pdfFiles.push(file);
-				}
-			}
-
-			if (imageFiles.length > 0 && onImagePaste) {
-				e.preventDefault();
-				onImagePaste(imageFiles);
-			}
-			if (pdfFiles.length > 0 && onPDFDrop) {
-				e.preventDefault();
-				onPDFDrop(pdfFiles);
-			}
-		};
-
-		// Primary: attach to container so it catches events bubbling from the textarea
-		const container = containerRef.current || divRef?.current;
-		if (container) {
-			container.addEventListener('paste', handlePaste);
-		}
-
-		// Fallback: document-level listener catches paste when chat area is visible but
-		// the textarea isn't focused (e.g. user focuses sidebar panel then Ctrl+V)
-		const handleDocumentPaste = (e: ClipboardEvent) => {
-			// Only intercept if the target is not already inside our container
-			// and our container is mounted in the DOM
-			const cont = containerRef.current || divRef?.current;
-			if (!cont) return;
-			if (cont.contains(e.target as Node)) return; // already handled by container listener
-			// Only fire if the paste comes from a non-input element (avoid stealing from other inputs)
-			const target = e.target as HTMLElement;
-			const tag = target?.tagName?.toLowerCase();
-			if (tag === 'input' || tag === 'textarea') return;
-
-			const items = Array.from(e.clipboardData?.items || []);
-			const imageFiles = items
-				.filter(i => i.type.startsWith('image/'))
-				.map(i => i.getAsFile())
-				.filter((f): f is File => f !== null);
-
-			if (imageFiles.length > 0 && onImagePaste) {
-				e.preventDefault();
-				onImagePaste(imageFiles);
-			}
-		};
-		document.addEventListener('paste', handleDocumentPaste);
-
-		return () => {
-			if (container) container.removeEventListener('paste', handlePaste);
-			document.removeEventListener('paste', handleDocumentPaste);
-		};
-	}, [divRef, onImagePaste, onPDFDrop]);
-
-	// Throttle drag over events to prevent jank
-	const lastDragOverTimeRef = React.useRef<number>(0);
-	const DRAG_THROTTLE_MS = 50; // Update at most every 50ms
-
-	// Handle drag and drop
-	const handleDragOver = React.useCallback((e: React.DragEvent) => {
-		e.preventDefault();
-		e.stopPropagation();
-
-		const now = Date.now();
-		if (now - lastDragOverTimeRef.current < DRAG_THROTTLE_MS) {
-			return;
-		}
-		lastDragOverTimeRef.current = now;
-
-		const hasFiles = Array.from(e.dataTransfer.items).some(item =>
-			item.type.startsWith('image/') || item.type === 'application/pdf'
-		);
-		if (hasFiles) {
-			setIsDragOver(true);
-		}
-	}, []);
-
-	const handleDragLeave = (e: React.DragEvent) => {
-		e.preventDefault();
-		e.stopPropagation();
-		setIsDragOver(false);
-	};
-
-	const handleDrop = (e: React.DragEvent) => {
-		e.preventDefault();
-		e.stopPropagation();
-		setIsDragOver(false);
-
-		const imageFiles = Array.from(e.dataTransfer.files).filter(file =>
-			file.type.startsWith('image/')
-		);
-		const pdfFiles = Array.from(e.dataTransfer.files).filter(file =>
-			file.type === 'application/pdf'
-		);
-
-		if (imageFiles.length > 0 && onImageDrop) {
-			onImageDrop(imageFiles);
-		}
-		if (pdfFiles.length > 0 && onPDFDrop) {
-			onPDFDrop(pdfFiles);
-		}
-	};
-
-	const handleImageUploadClick = () => {
-		imageInputRef.current?.click();
-	};
-
-	const handlePDFUploadClick = () => {
-		pdfInputRef.current?.click();
-	};
-
-	const handleImageInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-		const files = Array.from(e.target.files || []).filter(file =>
-			file.type.startsWith('image/')
-		);
-		if (files.length > 0 && onImageDrop) {
-			onImageDrop(files);
-		}
-		e.target.value = ''; // Reset input
-	};
-
-	const handlePDFInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-		const files = Array.from(e.target.files || []).filter(file =>
-			file.type === 'application/pdf'
-		);
-		if (files.length > 0 && onPDFDrop) {
-			onPDFDrop(files);
-		}
-		e.target.value = ''; // Reset input
-	};
-
-	return (
-		<div
-			ref={(node) => {
-				if (divRef) {
-					if (typeof divRef === 'function') {
-						divRef(node);
-					} else {
-						divRef.current = node;
-					}
-				}
-				containerRef.current = node;
-			}}
-			className={`
-				gap-x-1
-                flex flex-col p-2.5 relative input text-left shrink-0
-                rounded-2xl
-                bg-[#030304]
-				transition-all duration-200
-				border border-[rgba(255,255,255,0.08)] focus-within:border-[rgba(255,255,255,0.12)] hover:border-[rgba(255,255,255,0.12)]
-				${isDragOver ? 'border-blue-500 bg-blue-500/10' : ''}
-				max-h-[80vh] overflow-y-auto
-                ${className}
-            `}
-			onClick={(e) => {
-				onClickAnywhere?.()
-			}}
-			onDragOver={handleDragOver}
-			onDragLeave={handleDragLeave}
-			onDrop={handleDrop}
-		>
-			{/* Hidden file inputs - separate for images and PDFs */}
-			<input
-				ref={imageInputRef}
-				type="file"
-				accept="image/png,image/jpeg,image/webp,image/gif,image/svg+xml"
-				multiple
-				className="hidden"
-				onChange={handleImageInputChange}
-			/>
-			<input
-				ref={pdfInputRef}
-				type="file"
-				accept="application/pdf"
-				multiple
-				className="hidden"
-				onChange={handlePDFInputChange}
-			/>
-
-			{/* Image attachments section */}
-			{imageAttachments}
-
-			{/* PDF attachments section */}
-			{pdfAttachments}
-
-			{/* Selections section */}
-			{showSelections && selections && setSelections && (
-				<SelectedFiles
-					type='staging'
-					selections={selections}
-					setSelections={setSelections}
-					showProspectiveSelections={showProspectiveSelections}
-				/>
-			)}
-
-			{/* Input section - Modern Cursor-style layout */}
-			<div className="relative w-full flex items-end gap-2">
-				<div className="flex-1 min-w-0">
-					{children}
-				</div>
-
-				{/* Right-side icon bar - Cursor style */}
-				<div className="flex items-center gap-1 flex-shrink-0 pb-0.5">
-					{/* Image upload button */}
-					<button
-						type="button"
-						onClick={handleImageUploadClick}
-						className="flex-shrink-0 p-1.5 rounded hover:bg-void-bg-2-alt text-void-fg-4 hover:text-void-fg-2 transition-colors disabled:opacity-40 disabled:cursor-not-allowed"
-						aria-label="Upload images"
-						title="Upload images (or paste/drag & drop)"
-					>
-						<ImageIcon size={16} />
-					</button>
-
-					{/* PDF upload button */}
-					<button
-						type="button"
-						onClick={handlePDFUploadClick}
-						className="flex-shrink-0 p-1.5 rounded hover:bg-void-bg-2-alt text-void-fg-4 hover:text-void-fg-2 transition-colors disabled:opacity-40 disabled:cursor-not-allowed"
-						aria-label="Upload PDFs"
-						title="Upload PDFs (or paste/drag & drop)"
-					>
-						<FileText size={16} />
-					</button>
-
-					{/* Submit button */}
-					{isStreaming ? (
-						<ButtonStop onClick={onAbort} />
-					) : (
-						<ButtonSubmit
-							onClick={onSubmit}
-							disabled={isDisabled}
-						/>
-					)}
-				</div>
-
-				{/* Close button (X) if onClose is provided */}
-				{onClose && (
-					<div className='absolute -top-1 -right-1 cursor-pointer z-1'>
-						<IconX
-							size={12}
-							className="stroke-[2] opacity-80 text-void-fg-3 hover:brightness-95"
-							onClick={onClose}
-						/>
-					</div>
-				)}
-			</div>
-
-			{/* Bottom row - Model selector and settings */}
-			<div className='flex flex-row justify-between items-center gap-2 mt-1 pt-1 border-t border-void-border-3/50'>
-				{showModelDropdown && (
-					<div className='flex items-center flex-wrap gap-x-2 gap-y-1 text-nowrap flex-1 min-w-0'>
-						{featureName === 'Chat' && <ChatModeDropdown className='text-xs text-void-fg-3 bg-void-bg-1 border border-void-border-2 rounded py-0.5 px-1.5' />}
-						<ModelDropdown featureName={featureName} className='text-xs text-void-fg-3 bg-void-bg-1 rounded' />
-						<ReasoningOptionSlider featureName={featureName} />
-					</div>
-				)}
-
-				{/* Loading indicator */}
-				{isStreaming && loadingIcon && (
-					<div className="flex items-center">
-						{loadingIcon}
-					</div>
-				)}
-			</div>
-		</div>
-	);
-};
-
-
-
-
-type ButtonProps = ButtonHTMLAttributes<HTMLButtonElement>
-const DEFAULT_BUTTON_SIZE = 22;
-export const ButtonSubmit = ({ className, disabled, ...props }: ButtonProps & Required<Pick<ButtonProps, 'disabled'>>) => {
-
-	return <button
-		type='button'
-		className={`rounded-full flex-shrink-0 flex-grow-0 flex items-center justify-center
-			button-press-animation
-			${disabled ? 'bg-vscode-disabled-fg cursor-default opacity-50' : 'bg-white cursor-pointer hover:bg-gray-50'}
-			${className}
-		`}
-		disabled={disabled}
-		aria-label="Send message"
-		// data-tooltip-id='cortex-tooltip'
-		// data-tooltip-content={'Send'}
-		// data-tooltip-place='left'
-		{...props}
-	>
-		<IconArrowUp size={DEFAULT_BUTTON_SIZE} className="stroke-[2] p-[2px]" />
-	</button>
-}
-
-export const ButtonStop = ({ className, ...props }: ButtonHTMLAttributes<HTMLButtonElement>) => {
-	return <button
-		className={`rounded-full flex-shrink-0 flex-grow-0 cursor-pointer flex items-center justify-center
-			bg-white hover:bg-[var(--cortex-danger)]/5 button-press-animation
-			${className}
-		`}
-		type='button'
-		aria-label="Stop generation"
-		{...props}
-	>
-		<IconSquare size={DEFAULT_BUTTON_SIZE} className="stroke-[3] p-[7px] text-[var(--cortex-danger)]" />
-	</button>
-}
-
-
-
-const scrollToBottom = (divRef: { current: HTMLElement | null }) => {
-	if (divRef.current) {
-		divRef.current.scrollTop = divRef.current.scrollHeight;
-	}
-};
-
-
-
-const ScrollToBottomContainer = ({ children, className, style, scrollContainerRef }: { children: React.ReactNode, className?: string, style?: React.CSSProperties, scrollContainerRef: React.MutableRefObject<HTMLDivElement | null> }) => {
-	const [isAtBottom, setIsAtBottom] = useState(true); // Start at bottom
-
-	const divRef = scrollContainerRef
-
-	const onScroll = () => {
-		const div = divRef.current;
-		if (!div) return;
-
-		const isBottom = Math.abs(
-			div.scrollHeight - div.clientHeight - div.scrollTop
-		) < 4;
-
-		setIsAtBottom(isBottom);
-	};
-
-	// When children change (new messages added)
-	useEffect(() => {
-		if (isAtBottom) {
-			scrollToBottom(divRef);
-		}
-	}, [children, isAtBottom]); // Dependency on children to detect new messages
-
-	// Initial scroll to bottom
-	useEffect(() => {
-		scrollToBottom(divRef);
-	}, []);
-
-	return (
-		<div
-			ref={divRef}
-			onScroll={onScroll}
-			className={className}
-			style={style}
-		>
-			{children}
-		</div>
-	);
-};
-
-export const SelectedFiles = (
-	{ type, selections, setSelections, showProspectiveSelections, messageIdx, }:
-		| { type: 'past', selections: StagingSelectionItem[]; setSelections?: undefined, showProspectiveSelections?: undefined, messageIdx: number, }
-		| { type: 'staging', selections: StagingSelectionItem[]; setSelections: ((newSelections: StagingSelectionItem[]) => void), showProspectiveSelections?: boolean, messageIdx?: number }
-) => {
-
-	const accessor = useAccessor()
-	const commandService = accessor.get('ICommandService')
-	const modelReferenceService = accessor.get('ICortexideModelService')
-
-
-
-
-	// state for tracking prospective files
-	const { uri: currentURI } = useActiveURI()
-	const [recentUris, setRecentUris] = useState<URI[]>([])
-	const maxRecentUris = 10
-	const maxProspectiveFiles = 3
-	useEffect(() => { // handle recent files
-		if (!currentURI) return
-		setRecentUris(prev => {
-			const withoutCurrent = prev.filter(uri => uri.fsPath !== currentURI.fsPath) // remove duplicates
-			const withCurrent = [currentURI, ...withoutCurrent]
-			return withCurrent.slice(0, maxRecentUris)
-		})
-	}, [currentURI])
-	const [prospectiveSelections, setProspectiveSelections] = useState<StagingSelectionItem[]>([])
-
-
-	// handle prospective files
-	useEffect(() => {
-		const computeRecents = async () => {
-			const prospectiveURIs = recentUris
-				.filter(uri => !selections.find(s => s.type === 'File' && s.uri.fsPath === uri.fsPath))
-				.slice(0, maxProspectiveFiles)
-
-			const answer: StagingSelectionItem[] = []
-			for (const uri of prospectiveURIs) {
-				answer.push({
-					type: 'File',
-					uri: uri,
-					language: (await modelReferenceService.getModelSafe(uri)).model?.getLanguageId() || 'plaintext',
-					state: { wasAddedAsCurrentFile: false },
-				})
-			}
-			return answer
-		}
-
-		// add a prospective file if type === 'staging' and if the user is in a file, and if the file is not selected yet
-		if (type === 'staging' && showProspectiveSelections) {
-			computeRecents().then((a) => setProspectiveSelections(a))
-		}
-		else {
-			setProspectiveSelections([])
-		}
-	}, [recentUris, selections, type, showProspectiveSelections])
-
-
-	const allSelections = [...selections, ...prospectiveSelections]
-
-	if (allSelections.length === 0) {
-		return null
-	}
-
-	return (
-		<div className='flex items-center flex-wrap text-left relative gap-x-0.5 gap-y-1 pb-0.5'>
-
-			{allSelections.map((selection, i) => {
-
-				const isThisSelectionProspective = i > selections.length - 1
-
-				const thisKey = selection.type === 'CodeSelection' ? selection.type + selection.language + selection.range + selection.state.wasAddedAsCurrentFile + selection.uri.fsPath
-					: selection.type === 'File' ? selection.type + selection.language + selection.state.wasAddedAsCurrentFile + selection.uri.fsPath
-						: selection.type === 'Folder' ? selection.type + selection.language + selection.state + selection.uri.fsPath
-							: i
-
-				const SelectionIcon = (
-					selection.type === 'File' ? File
-						: selection.type === 'Folder' ? Folder
-							: selection.type === 'CodeSelection' ? Text
-								: (undefined as never)
-				)
-
-				return <div // container for summarybox and code
-					key={thisKey}
-					className={`flex flex-col space-y-[1px]`}
-				>
-					{/* tooltip for file path */}
-					<span className="truncate overflow-hidden text-ellipsis"
-						data-tooltip-id='cortex-tooltip'
-						data-tooltip-content={getRelative(selection.uri, accessor)}
-						data-tooltip-place='top'
-						data-tooltip-delay-show={3000}
-					>
-						{/* summarybox */}
-						<div
-							className={`
-								flex items-center gap-1 relative
-								px-1
-								w-fit h-fit
-								select-none
-								text-xs text-nowrap
-								border rounded-sm
-								${isThisSelectionProspective ? 'bg-void-bg-1 text-void-fg-3 opacity-80' : 'bg-void-bg-1 hover:brightness-95 text-void-fg-1'}
-								${isThisSelectionProspective
-									? 'border-void-border-2'
-									: 'border-void-border-1'
-								}
-								hover:border-void-border-1
-								transition-all duration-150
-							`}
-							onClick={() => {
-								if (type !== 'staging') return; // (never)
-								if (isThisSelectionProspective) { // add prospective selection to selections
-									setSelections([...selections, selection])
-								}
-								else if (selection.type === 'File') { // open files
-									voidOpenFileFn(selection.uri, accessor);
-
-									const wasAddedAsCurrentFile = selection.state.wasAddedAsCurrentFile
-									if (wasAddedAsCurrentFile) {
-										// make it so the file is added permanently, not just as the current file
-										const newSelection: StagingSelectionItem = { ...selection, state: { ...selection.state, wasAddedAsCurrentFile: false } }
-										setSelections([
-											...selections.slice(0, i),
-											newSelection,
-											...selections.slice(i + 1)
-										])
-									}
-								}
-								else if (selection.type === 'CodeSelection') {
-									voidOpenFileFn(selection.uri, accessor, selection.range);
-								}
-								else if (selection.type === 'Folder') {
-									// TODO!!! reveal in tree
-								}
-							}}
-						>
-							{<SelectionIcon size={10} />}
-
-							{ // file name and range
-								getBasename(selection.uri.fsPath)
-								+ (selection.type === 'CodeSelection' ? ` (${selection.range[0]}-${selection.range[1]})` : '')
-							}
-
-							{selection.type === 'File' && selection.state.wasAddedAsCurrentFile && messageIdx === undefined && currentURI?.fsPath === selection.uri.fsPath ?
-								<span className="text-[8px] void-opacity-60 text-void-fg-4">
-									{`(Current File)`}
-								</span>
-								: null
-							}
-
-							{type === 'staging' && !isThisSelectionProspective ? // X button
-								<div // box for making it easier to click
-									className='cursor-pointer z-1 self-stretch flex items-center justify-center'
-									onClick={(e) => {
-										e.stopPropagation(); // don't open/close selection
-										if (type !== 'staging') return;
-										setSelections([...selections.slice(0, i), ...selections.slice(i + 1)])
-									}}
-								>
-									<IconX
-										className='stroke-[2]'
-										size={10}
-									/>
-								</div>
-								: <></>
-							}
-						</div>
-					</span>
-				</div>
-
-			})}
-
-
-		</div>
-
-	)
-}
 
 
 type ToolHeaderParams = {
@@ -4574,24 +3810,6 @@ export const SidebarChat = () => {
 	const isLandingPage = previousMessages.length === 0
 
 
-	const initiallySuggestedPromptsHTML = <div className='flex flex-col gap-2 w-full text-nowrap text-void-fg-3 select-none'>
-		{[
-			'Summarize my codebase',
-			'How do types work in Rust?',
-			'Create a .voidrules file for me'
-		].map((text, index) => (
-			<div
-				key={index}
-				className='py-1 px-2 rounded text-sm bg-zinc-700/5 hover:bg-zinc-700/10 dark:bg-zinc-300/5 dark:hover:bg-zinc-300/10 cursor-pointer opacity-80 hover:opacity-100'
-				onClick={() => onSubmit(text)}
-			>
-				{text}
-			</div>
-		))}
-	</div>
-
-
-
 	const threadPageInput = <div key={'input' + chatThreadsState.currentThreadId}>
 		<div className='px-4'>
 			<CommandBarInChat />
@@ -4635,107 +3853,14 @@ export const SidebarChat = () => {
 		</div>
 	</div>
 
-    const keybindingService = accessor.get('IKeybindingService')
-    const quickActions: { id: string, label: string }[] = [
-        { id: 'void.explainCode', label: 'Explain' },
-        { id: 'void.refactorCode', label: 'Refactor' },
-        { id: 'void.addTests', label: 'Add Tests' },
-        { id: 'void.fixTests', label: 'Fix Tests' },
-        { id: 'void.writeDocstring', label: 'Docstring' },
-        { id: 'void.optimizeCode', label: 'Optimize' },
-        { id: 'void.debugCode', label: 'Debug' },
-    ]
-
-    const QuickActionsBar = () => (
-        <div className='w-full flex items-center justify-center gap-2 flex-wrap mt-3 select-none px-1'>
-            {quickActions.map(({ id, label }) => {
-                const kb = keybindingService.lookupKeybinding(id)?.getLabel()
-                return (
-                    <button
-                        key={id}
-                        className='px-3 py-1.5 rounded-full bg-gradient-to-br from-[var(--cortex-surface-2)] via-[var(--cortex-surface-3)] to-[var(--cortex-surface-4)] border border-void-border-3 text-xs text-void-fg-1 shadow-[0_3px_12px_rgba(0,0,0,0.45)] hover:-translate-y-0.5 transition-all duration-150 ease-out void-focus-ring'
-                        onClick={() => commandService.executeCommand(id)}
-                        title={kb ? `${label} (${kb})` : label}
-                    >
-                        <span>{label}</span>
-                        {kb && <span className='ml-1 px-1 rounded bg-[var(--vscode-keybindingLabel-background)] text-[var(--vscode-keybindingLabel-foreground)] border border-[var(--vscode-keybindingLabel-border)]'>{kb}</span>}
-                    </button>
-                )
-            })}
-        </div>
-    )
-
-    // Lightweight context chips: active file and model
-    const ContextChipsBar = () => {
-        const editorService = accessor.get('IEditorService')
-        const activeEditor = editorService?.activeEditor
-        // Try best-effort file label
-        const activeResource = activeEditor?.resource
-        const activeFileLabel = activeResource ? activeResource.path?.split('/').pop() : undefined
-        const modelSel = settingsState.modelSelectionOfFeature['Chat']
-        const modelLabel = modelSel ? `${modelSel.providerName}:${modelSel.modelName}` : undefined
-        if (!activeFileLabel && !modelLabel) return null
-        return (
-            <div className='w-full flex items-center gap-2 flex-wrap mt-2 mb-1 px-1'>
-                {activeFileLabel && (
-                    <span className='inline-flex items-center gap-1 px-2 py-0.5 rounded border border-void-border-3 bg-void-bg-1 text-void-fg-2 text-[11px]'>
-                        <span>File</span>
-                        <span className='text-void-fg-1'>{activeFileLabel}</span>
-                    </span>
-                )}
-                {modelLabel && (
-                    <span className='inline-flex items-center gap-1 px-2 py-0.5 rounded border border-void-border-3 bg-void-bg-1 text-void-fg-2 text-[11px]'>
-                        <span>Model</span>
-                        <span className='text-void-fg-1'>{modelLabel}</span>
-                    </span>
-                )}
-            </div>
-        )
-    }
-
-    const logoUri = useMemo(() => FileAccess.asBrowserUri('vs/workbench/browser/media/cortexide-main.png').toString(true), [])
-
-    const landingPageContent = <div
-		ref={sidebarRef}
-		className='w-full h-full max-h-full flex flex-col overflow-auto px-3'
-	>
-		{/* CortexIDE logo */}
-		<div className='flex flex-col items-center pt-6 pb-2 gap-1'>
-			<div className='w-20 h-20 rounded-full overflow-hidden' style={{ border: '2px solid var(--vscode-focusBorder, rgba(88,101,242,0.8))', boxShadow: '0 0 16px var(--vscode-focusBorder, rgba(88,101,242,0.4)), 0 4px 20px rgba(0,0,0,0.5)' }}>
-				<img
-					src={logoUri}
-					alt='CortexIDE'
-					className='w-full h-full object-cover'
-				/>
-			</div>
-			<span className='text-xs text-void-fg-3 tracking-widest uppercase'>CortexIDE</span>
-		</div>
-		<ErrorBoundary>
-			{landingPageInput}
-		</ErrorBoundary>
-
-		{/* Context chips */}
-		<ErrorBoundary>
-			<ContextChipsBar />
-		</ErrorBoundary>
-
-        {/* Quick Actions shortcuts */}
-        <ErrorBoundary>
-            <QuickActionsBar />
-        </ErrorBoundary>
-
-		{Object.keys(chatThreadsState.allThreads).length > 1 ? // show if there are threads
-			<ErrorBoundary>
-				<div className='pt-6 mb-2 text-void-fg-3 text-root select-none pointer-events-none'>{t('chat.previousThreads')}</div>
-				<PastThreadsList />
-			</ErrorBoundary>
-			:
-			<ErrorBoundary>
-				<div className='pt-6 mb-2 text-void-fg-3 text-root select-none pointer-events-none'>{t('chat.suggestions')}</div>
-				{initiallySuggestedPromptsHTML}
-			</ErrorBoundary>
-		}
-	</div>
+	const landingPageContent = <LandingPage
+		sidebarRef={sidebarRef}
+		inputSection={landingPageInput}
+		showPreviousThreads={Object.keys(chatThreadsState.allThreads).length > 1}
+		suggestionsLabel={t('chat.suggestions')}
+		previousThreadsLabel={t('chat.previousThreads')}
+		onSubmitPrompt={onSubmit}
+	/>
 
 
 	// const threadPageContent = <div>
