+++ src/vs/workbench/contrib/cortexide/browser/react/src/sidebar-tsx/composer/ScrollToBottomContainer.tsx
@@ -1,0 +1,50 @@
+/*--------------------------------------------------------------------------------------
+ *  Copyright 2025 Glass Devtools, Inc. All rights reserved.
+ *  Licensed under the Apache License, Version 2.0. See LICENSE.txt for more information.
+ *--------------------------------------------------------------------------------------*/
+
+import React, { useEffect, useState } from 'react';
+
+const scrollToBottom = (divRef: { current: HTMLElement | null }) => {
+	if (divRef.current) {
+		divRef.current.scrollTop = divRef.current.scrollHeight;
+	}
+};
+
+export const ScrollToBottomContainer = ({ children, className, style, scrollContainerRef }: { children: React.ReactNode, className?: string, style?: React.CSSProperties, scrollContainerRef: React.MutableRefObject<HTMLDivElement | null> }) => {
+	const [isAtBottom, setIsAtBottom] = useState(true);
+
+	const divRef = scrollContainerRef;
+
+	const onScroll = () => {
+		const div = divRef.current;
+		if (!div) return;
+
+		const isBottom = Math.abs(
+			div.scrollHeight - div.clientHeight - div.scrollTop
+		) < 4;
+
+		setIsAtBottom(isBottom);
+	};
+
+	useEffect(() => {
+		if (isAtBottom) {
+			scrollToBottom(divRef);
+		}
+	}, [children, isAtBottom, divRef]);
+
+	useEffect(() => {
+		scrollToBottom(divRef);
+	}, [divRef]);
+
+	return (
+		<div
+			ref={divRef}
+			onScroll={onScroll}
+			className={className}
+			style={style}
+		>
+			{children}
+		</div>
+	);
+};
